+++ src/criticality.rs
@@ -27,7 +27,7 @@ pub enum Criticality {
 
 impl Display for Criticality {
     #[allow(clippy::use_debug)]
-    fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
         write!(f, "{}", format!("{:?}", self).to_lowercase())
     }
 }