+++ src/utils.rs
@@ -165,7 +165,7 @@ impl Benchmark {
 }
 
 impl fmt::Display for Benchmark {
-    fn fmt(&self, f: &mut fmt::Formatter) -> Result<(), fmt::Error> {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> Result<(), fmt::Error> {
         write!(
             f,
             "{}: {}.{}s",